webapp/requirements.txt
streamlit==1.28.0
hopsworks[python]==4.1.5
confluent-kafka>=2.3.0
pyarrow>=14.0.1
pandas==2.1.4
xgboost==1.7.3
matplotlib==3.7.2
python-dotenv==1.0.0
requests==2.31.0